backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
argon2-cffi==23.1.0
bcrypt==4.1.1
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0